# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: MailBrief API
  description: |
    REST API for MailBrief, a multi-user email digest dashboard.
    Users receive AI-generated daily digests of their newsletters via managed inboxes.
  version: 1.0.0
  contact:
    name: MailBrief
    url: https://mailbrief.io
servers:
  - url: https://mailbrief.io
    description: Production
  - url: http://localhost:8080
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /webhook/email:
    post:
      operationId: ingestEmail
      summary: Ingest a raw email
      description: |
        Receives a raw RFC 5322 email from the Cloudflare Email Worker.
        The X-Original-To header identifies the target managed inbox.
      tags: [Webhooks]
      security:
        - WebhookSecret: []
      parameters:
        - name: X-Original-To
          in: header
          required: true
          schema:
            type: string
          description: The managed inbox address the email was sent to.
      requestBody:
        required: true
        content:
          message/rfc822:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Email accepted for processing.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /cron/digests:
    post:
      operationId: triggerDigests
      summary: Trigger digest generation
      description: Fires the digest generation cron job for all eligible users.
      tags: [Webhooks]
      security:
        - ApiKey: []
      responses:
        "200":
          description: Digest generation triggered.
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    example: triggered

  # ── Accounts ────────────────────────────────────────────────────────
  /api/accounts:
    get:
      operationId: listAccounts
      summary: List email accounts
      description: Returns all managed email accounts for the authenticated user.
      tags: [Accounts]
      responses:
        "200":
          description: List of email accounts.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailAccountList"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createAccount
      summary: Create an email account
      description: Provisions a new managed inbox with the given username.
      tags: [Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [Username]
              properties:
                Username:
                  type: string
                  minLength: 3
                  maxLength: 30
                  pattern: "^[a-z0-9][a-z0-9._-]*[a-z0-9]$"
      responses:
        "201":
          description: Account created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailAccount"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Username already taken.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/accounts/check-username:
    get:
      operationId: checkUsername
      summary: Check username availability
      description: Validates whether a username is available and well-formed.
      tags: [Accounts]
      parameters:
        - name: username
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Availability result.
          content:
            application/json:
              schema:
                type: object
                required: [available]
                properties:
                  available:
                    type: boolean
                  reason:
                    type: string
  /api/accounts/{id}:
    delete:
      operationId: deleteAccount
      summary: Delete an email account
      description: Permanently removes a managed email account.
      tags: [Accounts]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "204":
          description: Account deleted.
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Emails ──────────────────────────────────────────────────────────
  /api/emails:
    get:
      operationId: listEmails
      summary: List emails
      description: Returns a paginated, filterable list of emails.
      tags: [Emails]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 200
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: is_read
          in: query
          schema:
            type: boolean
        - name: is_starred
          in: query
          schema:
            type: boolean
        - name: is_archived
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
          description: Full-text search across subject and body.
        - name: topic
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: from_address
          in: query
          schema:
            type: string
        - name: label
          in: query
          schema:
            type: string
          description: Filter by label ID.
        - name: has_attachment
          in: query
          schema:
            type: boolean
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, oldest, relevance]
            default: newest
        - name: since
          in: query
          schema:
            type: string
            format: date-time
        - name: before
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Paginated email list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailList"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/emails/{id}:
    get:
      operationId: getEmail
      summary: Get email details
      description: Returns the full email with extraction results, attachments, and thread count.
      tags: [Emails]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Email with enrichments.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateEmail
      summary: Update email state
      description: Toggles read, starred, or archived status on a single email.
      tags: [Emails]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailState"
      responses:
        "200":
          description: Email updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteEmail
      summary: Delete an email
      description: Permanently removes an email and its extraction.
      tags: [Emails]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Email deleted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: deleted
  /api/emails/{id}/thread:
    get:
      operationId: getEmailThread
      summary: Get email thread
      description: Returns all emails in the same conversation thread.
      tags: [Emails]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Thread emails ordered chronologically.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ThreadEmail"
  /api/emails/{id}/reprocess:
    post:
      operationId: reprocessEmail
      summary: Reprocess an email
      description: Queues the email for re-extraction by the LLM pipeline.
      tags: [Emails]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Email queued for reprocessing.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
  /api/emails/{id}/unsubscribe:
    post:
      operationId: unsubscribeEmail
      summary: Unsubscribe from sender
      description: |
        Attempts to unsubscribe using the List-Unsubscribe header from this email.
        Returns the method used (url or mailto) and the target.
      tags: [Emails]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Unsubscribe result.
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                  method:
                    type: string
                    enum: [url, mailto]
                  url:
                    type: string
                  address:
                    type: string
  /api/emails/batch:
    patch:
      operationId: batchUpdateEmails
      summary: Batch update email state
      description: Applies the same state change to multiple emails at once.
      tags: [Emails]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids, state]
              properties:
                ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                state:
                  $ref: "#/components/schemas/EmailState"
      responses:
        "200":
          description: Emails updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
    delete:
      operationId: batchDeleteEmails
      summary: Batch delete emails
      description: Permanently removes multiple emails.
      tags: [Emails]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              properties:
                ids:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        "200":
          description: Emails deleted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: deleted

  # ── Email Feedback ──────────────────────────────────────────────────
  /api/emails/{id}/feedback:
    post:
      operationId: createEmailFeedback
      summary: Submit email feedback
      description: Rate an email extraction as useful or not useful for quality improvement.
      tags: [Feedback]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [Rating]
              properties:
                Rating:
                  type: string
                  enum: [useful, not_useful]
      responses:
        "200":
          description: Feedback recorded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailFeedback"
    get:
      operationId: getEmailFeedback
      summary: Get email feedback
      description: Returns existing feedback for an email, if any.
      tags: [Feedback]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Feedback record.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailFeedback"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteEmailFeedback
      summary: Remove email feedback
      description: Deletes previously submitted feedback for an email.
      tags: [Feedback]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Feedback deleted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: deleted

  # ── Email Labels ────────────────────────────────────────────────────
  /api/emails/{id}/labels:
    get:
      operationId: getEmailLabels
      summary: Get labels for an email
      description: Returns all labels applied to a specific email.
      tags: [Labels]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Labels for the email.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Label"
  /api/emails/{id}/labels/{labelId}:
    put:
      operationId: addLabelToEmail
      summary: Apply a label to an email
      description: Associates an existing label with an email.
      tags: [Labels]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
        - name: labelId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Label applied.
    delete:
      operationId: removeLabelFromEmail
      summary: Remove a label from an email
      description: Disassociates a label from an email.
      tags: [Labels]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
        - name: labelId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Label removed.
  /api/emails/labels/batch-add:
    post:
      operationId: batchAddLabel
      summary: Batch add label to emails
      description: Applies a single label to multiple emails at once.
      tags: [Labels]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [EmailIDs, LabelID]
              properties:
                EmailIDs:
                  type: array
                  items:
                    type: string
                    format: uuid
                LabelID:
                  type: string
                  format: uuid
      responses:
        "204":
          description: Label applied to all specified emails.
  /api/emails/labels/batch-remove:
    post:
      operationId: batchRemoveLabel
      summary: Batch remove label from emails
      description: Removes a single label from multiple emails at once.
      tags: [Labels]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [EmailIDs, LabelID]
              properties:
                EmailIDs:
                  type: array
                  items:
                    type: string
                    format: uuid
                LabelID:
                  type: string
                  format: uuid
      responses:
        "204":
          description: Label removed from all specified emails.

  # ── Subscriptions ───────────────────────────────────────────────────
...
  /api/subscriptions/unsubscribe:
    post:
      operationId: batchUnsubscribe
      summary: Batch unsubscribe from senders
      description: Marks one or more senders as unsubscribed.
      tags: [Subscriptions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [FromAddresses]
              properties:
                FromAddresses:
                  type: array
                  items:
                    type: string
                    format: email
      responses:
        "200":
          description: Senders unsubscribed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"

  # ── Digests ─────────────────────────────────────────────────────────
  /api/digests:
    get:
      operationId: listDigests
      summary: List digests
      description: Returns a paginated list of generated digests.
      tags: [Digests]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        "200":
          description: Paginated digest list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DigestList"
  /api/digests/{id}:
    get:
      operationId: getDigest
      summary: Get digest details
      description: Returns a single digest with its items.
      tags: [Digests]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Digest with items.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Digest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteDigest
      summary: Delete a digest
      description: Permanently removes a digest.
      tags: [Digests]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Digest deleted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: deleted
  /api/digests/{id}/share:
    post:
      operationId: shareDigest
      summary: Create a share link
      description: Generates a public share token for a digest.
      tags: [Digests]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Share token created.
          content:
            application/json:
              schema:
                type: object
                required: [ShareToken]
                properties:
                  ShareToken:
                    type: string
    delete:
      operationId: unshareDigest
      summary: Revoke share link
      description: Removes the public share token from a digest.
      tags: [Digests]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Share revoked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
  /api/digests/preview:
    post:
      operationId: previewDigest
      summary: Preview digest contents
      description: Shows which emails would be included in a digest for the given time window, without generating it.
      tags: [Digests]
      parameters:
        - name: period_start
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: period_end
          in: query
          required: true
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Digest preview.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DigestPreview"
  /api/digests/generate:
    post:
      operationId: generateDigest
      summary: Generate a digest on demand
      description: Creates a new digest for the specified time window using the LLM pipeline.
      tags: [Digests]
      parameters:
        - name: period_start
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: period_end
          in: query
          required: true
          schema:
            type: string
            format: date-time
      responses:
        "201":
          description: Digest generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Digest"
        "409":
          description: A digest already exists for this period.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: No processed emails in the given period.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Session token from Better Auth (passed as Authorization header).
    WebhookSecret:
      type: apiKey
      in: header
      name: X-Webhook-Secret
      description: Shared secret for Cloudflare Email Worker webhook.
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for cron and service-to-service calls.
...
  responses:
    BadRequest:
      description: Invalid request body or parameters.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid authentication.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    StatusResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          example: ok

    # ── Email ───────────────────────────────────────────────────────
...
    EmailAccountList:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/EmailAccount"
        total:
          type: integer

    # ── Digest ──────────────────────────────────────────────────────
...